+++ src/components/payment/index.ts
@@ -4,3 +4,4 @@ export * from 'components/payment/UserAddress'
 export * from 'components/payment/PriceInfo'
 export * from 'components/payment/PaymentMethods'
 export * from 'components/payment/Confirmation'
+export * from 'components/payment/DaumPostCode'
