+++ src/styles/components/payment/UserAddress.module.scss
@@ -9,10 +9,6 @@
     font-weight: 700;
     padding-bottom: 20px;
   }
-  // .receiver{}
-  // .address{}
-  // .mobile{}
-  // .email{}
   label {
     display: flex;
     span {
